refactor(close-validation): split lib/close-validation.js monolith into close-validation/ modules (refs #3994)
